+++ include/fuse_common.h
@@ -90,7 +90,6 @@ struct fuse_file_info {
  * FUSE_CAP_POSIX_LOCKS: filesystem supports "remote" locking
  * FUSE_CAP_ATOMIC_O_TRUNC: filesystem handles the O_TRUNC open flag
  * FUSE_CAP_EXPORT_SUPPORT: filesystem handles lookups of "." and ".."
- * FUSE_CAP_BIG_WRITES: filesystem can handle write size larger than 4kB
  * FUSE_CAP_DONT_MASK: don't apply umask to file mode on create operations
  * FUSE_CAP_SPLICE_WRITE: ability to use splice() to write to the fuse device
  * FUSE_CAP_SPLICE_MOVE: ability to move data to the fuse device with splice()
@@ -107,7 +106,8 @@ struct fuse_file_info {
 #define FUSE_CAP_POSIX_LOCKS		(1 << 1)
 #define FUSE_CAP_ATOMIC_O_TRUNC		(1 << 3)
 #define FUSE_CAP_EXPORT_SUPPORT		(1 << 4)
-#define FUSE_CAP_BIG_WRITES		(1 << 5)
+/* (1 << 5) used to be FUSE_CAP_BIG_WRITES, which is now
+   always enabled */
 #define FUSE_CAP_DONT_MASK		(1 << 6)
 #define FUSE_CAP_SPLICE_WRITE		(1 << 7)
 #define FUSE_CAP_SPLICE_MOVE		(1 << 8)